Adjust init of OAuth2XSUAAEnricher so that this survives an invalid config
Reason: this change is in the context of providing a configuration validator for the auth config.
The valiation happens **after** initializing the enricher. So an invalid config would not be
detected by the validation, but here during that init.
In order to get the validators detect the mis-configuration we do not fail here.

The disavantage is that the field tokenClient is not final anymore.

Maybe it would be better to work with a failure here in that class and to see that
as some kind of validation rather than using a real validator and prevent an exception
raised here from bubbling up.

Other approach: ensure that the validators are executed earlier ...